% File man/describeCorrelation.Rd
% Part of the EASI package, https://cwendorf.github.io/EASI

\name{describeCorrelation}
\alias{describeCorrelation}
\alias{descCorrelation.default}
\title{Describe Correlation}
\description{Provides a square matrix/frame of correlations for a data set.}

\usage{
\method{describeCorrelation}{default}(\dots,digits=3)
}

\arguments{
\item{\dots}{a set of (non-empty) numeric vectors of data values.}
\item{digits}{the specified number of decimal places.}
}

\details{
Tutorials and examples from published work are available at \url{https://cwendorf.github.io/EASI/}
}