% File man/EASI-package.Rd
% Part of the EASI package, https://cwendorf.github.io/EASI

\name{EASI-package}
\alias{EASI-package}
\alias{EASI}
\docType{package}

\title{Estimation Approach to Statistical Inference}
\description{
EASI implements features of estimation statistics (or the "new statistics"). Included functions calculate, test, and plot confidence intervals for means, mean differences, and mean contrasts for between- and within-subjects single-factor, factorial, and mixed designs and can take either raw data or summary statistics as input. Additional functions calculate the standardized effect sizes and statistical significance tests associated with comparions among the groups and variables.
}

\author{
Craig A. Wendorf\cr 
Department of Psychology\cr 
University of Wisconsin - Stevens Point\cr 
\email{cwendorf@uwsp.edu} 
} 

\references{
Wendorf, C.A. (2019). EASI: Estimation Approach to Statistical Inference [R Package]. https://cwendorf.github.io/EASI/
}
