% File man/EASI-package.Rd
% Part of the EASI package, https://cwendorf.github.io/EASI

\name{EASI-package}
\alias{EASI-package}
\alias{EASI}
\docType{package}

\title{Estimation Approach to Statistical Inference}
\description{
\code{EASI} implements features of estimation statistics. Its primary functions calculate, test, and plot confidence intervals for means, mean differences, and mean contrasts for between- and within-subjects single-factor, factorial, and mixed designs. Additional functions calculate the standardized effect sizes and statistical significance tests associated with comparisons among the groups and variables. The functions can take either raw data or summary statistics as input.
}

\details{
\tabular{ll}{
Package: \tab EASI\cr
Type: \tab Package\cr
}
A full description of EASI, along with complete tutorials and examples from published work, is available on the corresponding web site:
\url{https://cwendorf.github.io/EASI/}
}

\author{
Craig A. Wendorf\cr 
Department of Psychology\cr 
University of Wisconsin - Stevens Point\cr 
\email{cwendorf@uwsp.edu} 
} 

\references{
Wendorf, C.A. (2019). EASI: Estimation Approach to Statistical Inference [R Package]. \url{https://cwendorf.github.io/EASI/}
}
